Rewrite lightHandler, improve light saving format
Added a bit of documentation, mostly for myself because my brain has not a lot of differences to a noodle sieve.
The entirety of the lightHandler now seems to make a lot more sense (I really hope that's not just in my head, and it's still terrible).
Also, the file storing the lights can now hold more than just one lamp and the overall information saved got more extensive. The whole file makes more sense now, too (the note from the lightHandler also applies here).
Lastly, I remembered that JS can be quite the pain sometimes and questioned my life several times in a rather short period of time, so nothing out of the ordinary.